requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.1
django-filter==1.1.0
django-rest-framework==0.1.0
djangorestframework==3.7.7
httpie==0.9.9
idna==2.6
Pygments==2.2.0
pytz==2017.3
requests==2.18.4
urllib3==1.22
